Allow tabs to not need href, move empty-state to holocene, make dropdown button thinner
