+++ src/lib/components/event/event-empty-row.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import EmptyState from '$lib/components/empty-state.svelte';
+  import EmptyState from '$lib/holocene/empty-state.svelte';
 </script>
 
 <article class="row">
