+++ src/lib/pages/workflow-pending-activities.svelte
@@ -4,7 +4,7 @@
 
   import Icon from '$holocene/icon/index.svelte';
   import Badge from '$holocene/badge.svelte';
-  import EmptyState from '$lib/components/empty-state.svelte';
+  import EmptyState from '$lib/holocene/empty-state.svelte';
   import Link from '$lib/components/link.svelte';
   import CodeBlock from '$lib/holocene/code-block.svelte';
   import {
