+++ src/lib/pages/workflow-stack-trace.svelte
@@ -7,7 +7,7 @@
 
   import CodeBlock from '$lib/holocene/code-block.svelte';
   import Button from '$holocene/button.svelte';
-  import EmptyState from '$lib/components/empty-state.svelte';
+  import EmptyState from '$lib/holocene/empty-state.svelte';
   import PageTitle from '$lib/holocene/page-title.svelte';
   import Loading from '$lib/holocene/loading.svelte';
   import Link from '$lib/components/link.svelte';
