+++ src/routes/namespaces/[namespace]/archival/index@root.svelte
@@ -60,7 +60,7 @@
   import WorkflowsSummaryRow from '$lib/components/workflow/workflows-summary-row.svelte';
   import WorkflowFilters from './_workflow-filters.svelte';
   import Pagination from '$holocene/pagination.svelte';
-  import EmptyState from '$lib/components/empty-state.svelte';
+  import EmptyState from '$lib/holocene/empty-state.svelte';
   import CodeBlock from '$lib/holocene/code-block.svelte';
   import { timeFormat } from '$lib/stores/time-format';
   import PageTitle from '$lib/holocene/page-title.svelte';
